% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload_package.R
\name{auto_detect_package_info}
\alias{auto_detect_package_info}
\title{Auto-detect package information from current project}
\usage{
auto_detect_package_info(verbose = FALSE)
}
\description{
Auto-detect package information from current project
}
\keyword{internal}
