% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload_package.R
\name{reinstall_from_github}
\alias{reinstall_from_github}
\title{Reinstall package from GitHub}
\usage{
reinstall_from_github(repo, branch, verbose = FALSE)
}
\description{
Reinstall package from GitHub
}
\keyword{internal}
