% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload_package.R
\name{safe_uninstall_package}
\alias{safe_uninstall_package}
\title{Safely uninstall package}
\usage{
safe_uninstall_package(package, verbose = FALSE)
}
\description{
Safely uninstall package
}
\keyword{internal}
